Move Stanford paver files to Stanford directory
