+++ pavelib/__init__.py
@@ -2,4 +2,4 @@
 paver commands
 """
 from . import assets, servers, docs, prereqs, quality, tests, js_test, i18n, bok_choy, acceptance_test
-from . import stanford_i18n
+from openedx.stanford.pavelib import i18n as stanford_i18n
